pyproject.toml
[project]
name = "nutrilens-ml"
version = "0.0.1"
description = "ML pipeline for NutriLens — plate scanning, pour volume estimation, follow-up logic"
requires-python = ">=3.11,<3.13"
readme = "README.md"
license = { text = "Proprietary" }
dependencies = [
    # Core
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0",
    # Data
    "numpy>=1.26",
    "pillow>=10.2",
    "boto3>=1.34",
    # ML
    "torch>=2.3",
    "torchvision>=0.18",
    # Inference & export
    "onnx>=1.16",
    "onnxruntime>=1.18",
    # Metrics & calibration
    "scikit-learn>=1.4",
]

[project.optional-dependencies]
serve = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
]
export = [
    # coremltools is macOS-first; install the extra on the Mac used for export
    "coremltools>=7.2",
]
track = [
    "wandb>=0.17",
]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-pyyaml",
]

[project.scripts]
nutrilens-ml = "nutrilens_ml.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nutrilens_ml"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
explicit_package_bases = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["wandb.*", "coremltools.*", "onnxruntime.*", "torchvision.*", "boto3.*", "botocore.*", "sklearn.*", "cv2.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
